Create a transcription configuration object
The idea is to have an object that can be passed (in part) to the
subtasks of audio conversion, diarization, and transcription. I keep out
everything that is related to the state of the GUI.

I am using a Munch object, which behaves like a fancy dict whos values can
be accessed using the dot syntax — like attributes of an object — and can
be easily serialized into a YAML file.